Move openxr to headset.c;
